Convert IdleDelay struct into RandomizedDelay class
