+++ src/fheroes2/battle/battle_animation.cpp
@@ -23,6 +23,25 @@
 #include "settings.h"
 #include <algorithm>
 
+RandomizedDelay::RandomizedDelay( uint32_t delay )
+    : TimeDelay( delay )
+    , halfDelay( delay / 2 )
+    , timerIsSet( false )
+{}
+
+bool RandomizedDelay::checkDelay()
+{
+    if ( !timerIsSet ) {
+        // Randomize delay as 0.75 to 1.25 original value
+        second = Rand::Get( 0, halfDelay ) + halfDelay * 3 / 2;
+        timerIsSet = true;
+    }
+    const bool res = Trigger();
+    if ( res )
+        timerIsSet = false;
+    return res;
+}
+
 AnimationSequence::AnimationSequence( const std::vector<int> & seq )
     : _seq( seq )
     , _currentFrame( 0 )
