+++ src/fheroes2/battle/battle_animation.h
@@ -23,7 +23,21 @@
 
 #include "bin_info.h"
 #include "gamedefs.h"
+#include "game_delays.h"
 #include "monster_info.h"
+
+// This timer is used for randomized idle animation delays, automatically setting it in range of 75%-125% of the intended value
+class RandomizedDelay : TimeDelay
+{
+    uint32_t halfDelay;
+    bool timerIsSet;
+
+public:
+    RandomizedDelay( uint32_t delay );
+
+    // This function triggers the current delay, returning true if it is passed and automatically resets the timer.
+    bool checkDelay();
+};
 
 struct monsterReturnAnim
 {
