+++ src/fheroes2/game/game_delays.cpp
@@ -59,25 +59,6 @@ bool TimeDelay::Trigger( uint32_t customDelay )
     return true;
 }
 
-IdleTimer::IdleTimer( uint32_t delay )
-    : TimeDelay( delay )
-    , halfDelay( delay / 2 )
-    , timerIsSet( false )
-{}
-
-bool IdleTimer::checkIdleDelay()
-{
-    if ( !timerIsSet ) {
-        // Randomize delay as 0.75 to 1.25 original value
-        second = Rand::Get( 0, halfDelay ) + halfDelay * 3 / 2;
-        timerIsSet = true;
-    }
-    const bool res = Trigger();
-    if ( res )
-        timerIsSet = false;
-    return res;
-}
-
 namespace Game
 {
     void AnimateDelaysInitialize( void );
