+++ src/fheroes2/game/game_delays.h
@@ -31,13 +31,4 @@ struct TimeDelay : std::pair<SDL::Time, uint32_t>
     void Reset( void );
     bool Trigger( uint32_t customDelay = 0 );
 };
-
-struct IdleTimer : TimeDelay
-{
-    uint32_t halfDelay;
-    bool timerIsSet;
-
-    IdleTimer( uint32_t delay );
-    bool checkIdleDelay();
-};
 #endif
